pyproject.toml
[project]
name = "gitlab-notifier"
version = "0.1.0"
description = "Telegram notifications for self-hosted GitLab"
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot[ext]>=21.0",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "alembic>=1.13",
  "asyncpg>=0.29",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "cryptography>=42",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "aiosqlite>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gitlab_notifier"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
